+++ src/Effects/Fire.h
@@ -8,7 +8,7 @@ class EffectFire: public Effect{
     EffectFire();
     ~EffectFire();
   
-    void proceed(CRGB *leds, int numLeds); 
+    void proceed(CRGB *leds, uint16_t numLeds); 
     void reset();
 };
 
@@ -27,10 +27,10 @@ inline void EffectFire::reset(){
 #define FIRE_X_SCALE 20 
 #define FIRE_Y_SCALE 3
 
-inline void EffectFire::proceed(CRGB *leds, int numLeds){  
+inline void EffectFire::proceed(CRGB *leds, uint16_t numLeds){  
   CRGBPalette16 pal = LavaColors_p;
   
-  for(int i = 0; i < numLeds ; i++) {        
+  for(uint16_t i = 0; i < numLeds ; i++) {        
     leds[i] = ColorFromPalette(pal, 
                               min( (i % 50) * (inoise8((i % 50) * FIRE_X_SCALE, millis() * FIRE_Y_SCALE * numLeds / 255))  >> 6, 255),
                               255,
